Revert "adding enum roletype"
This reverts commit 93b251e751ff0ac8c0d8b7b10d881923bcf81da0.

+++ BIS.API/Controller/RoleController.cs
@@ -16,8 +16,8 @@ namespace BIS.API.Controller
         [HttpGet]
         public IActionResult GetAll()
         {
-            int CorpsId = HttpContext.GetCorpsId();
-            int DivisionId = HttpContext.GetDivisionId();
+            long CorpsId = HttpContext.GetCorpsId();
+            long DivisionId = HttpContext.GetDivisionId();
             return Ok(_roleManager.GetAll(CorpsId, DivisionId));
         }
         [HttpPost]
